Add proper prebuilts to actually install "asan-test".
Because asan_test depends on internal implementation details of ASan, so
we must use prebuilts whenever we update the platform's compiler. This
CL properly installs these new prebuilts as "asan-test" (note the
hyphen). When Soong has support for conditional builds, we can switch
this back to the underscore version, but right now this causes conflicts
with the Android.bp build rules for asan_test in compiler-rt.

Test: Build for all 6 ABI architectures, and verified proper prebuilts
are copied.

Change-Id: Ie85b37a1216d28ac212bf2c123b2f84bf1f1bffe