Support a prebuilt libFuzzer.a for use in Android target builds.
Bug: http://b/35432050

This change adds prebuilt rules for libFuzzer.a for each Android target
architecture. libFuzzer.a needs to be a prebuilt, because it must match
the host Clang compiler that is used to compile other code with it. We
previously were always "lucky" that our host Clang remained mostly
compatible with the on-device LLVM sources (which usually trail behind
the host compiler by a few weeks).

For expediency, we use a fixed version number for the host Clang
prebuilt to match the libFuzzer.a prebuilts. This requires that we
update yet another version number when we upgrade our toolchain, but it
is faster than generating a new set of rules in Soong for creating
"fuzz_test" executables.

Test: Builds vts fuzzer tests successfully.

(cherry picked from commit b705d67f188995750344296e80d67776e0252cc2)

Change-Id: Ic5bcfa95643a7e008fefef288813e0e0b5b8b8a0