Mark libclang_rt* prebuilts as VNDK
There was a bug in the build system that allowed a VNDK lib to link
against to a lib in vendor partition. These libclang_rt* prebuilts
are used by some VNDK libraries (e.g. libstagefright_foundatation) which
are sanitized.

Since the bug in the build system has been fixed, we need to fix the
violation by correctly marking these prebuilt as VNDK so that they are
installed to the system partition.

Bug: 64994918
Test: no link_type check error on libclang_rt when building 2017 pixel
devices where BOARD_VNDK_VERSION=current.

Change-Id: Ib5a831734c9d40387153a5e07791bbbf3b93f653