Add Darwin host libraries to filegroup only if present
Bug: http://b/64121881

The Darwin prebuilts are not always available - for instance on the
Linux targets in go/ab/aosp-master-plus-llvm  or when testing a new
compiler.  Add the Darwin host LLVM libraries to the corresponding file
group only if they are present.  This makes the missing Darwin libraries
result in failure iff they are actually needed.

Test: Test new compiler update and test_compiler.py in
aosp-master-plus-llvm branch.

Change-Id: Ib8b5f869f34f63cee452a7553d27d1a0bb36ee5e