README: Fix Linux kernel links
The 4.9 link does not work because the android-4.9 branch was moved to
deprecated. Use the android-4.9-q branch.

The 4.14 link works but it does not show the clang version after the
build.config refactoring; use the build.config.common for all three
branches because that change was added to all of them:

https://android.googlesource.com/kernel/common/+/7d642373db4c6c1c6ed8af02492047896beb4a47

The 4.19 link does not work because the cuttlefish build.config was
removed for GKI:

https://android.googlesource.com/kernel/common/+/3f0d9e29849d2ab72600269041fce18ff1b0b99a

Change-Id: Iac0c78ccc444cf2c70a50d15e55c08b945aceb65
Reported-by: Souhrid Nandi <snkm1999@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>