libclang_rt.*-aarch64-android libraries are used for host as well
This change adds 'host_supported: true' for all
libclang_rt.*-aarch64-android libraries so that they can be used also
for the ARM64-based host targets. Although they have the "-android"
suffix which means they are built for Android, it actually means
"-bionic". The new ARM64-based host target 'linux_cross' is using Bionic
instead of glibc. So, these libraries are configured to be available
there, otherwise we will have to create a whole new set of libraries for
the new target.

Bug: 159685774
Test: HOST_CROSS_OS=linux_cross m

Change-Id: I2de6567c4a00ff18d090854d06978f0ea6d344da